Disable northside "Johnson Trigger" region at start
This region at the east of Northside should be disabled at start similar to the one in Redmond.
It is conditionally enabled by the "Mr. Johnson" trigger which also spawns the actors.
Entering it starts a conversation with Mr. Johnson, which increments the run step.
That creates wacky behaviour when one walks into it during other runs.